Split legacy Node test monolith
